build: make rsxs configurable
Also, there were a few X libs (xt and xmu) that appear to only be used by
rsxs. We still check for them and link to them so that our packaging scripts
(Ubuntu) are able to resolve deps correctly, otherwise the linking in the XBMC
binary is unnecessary.

We might consider removing these links and hinting to packagers instead.